requirements.txt
scipy>=1.1.0
SoundFile>=0.10.2
sacred==0.7.4
librosa>=0.6.2
numpy>=1.15.0
tqdm>=4.28.1
mir-eval==0.7
# git+https://github.com/craffel/mir_eval.git
mido>=1.2.9
Pillow>=6.2.0
dtw-python==1.3.0
einops==0.6.1
torch==2.0.1
PyYAML==6.0